engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
elevenlabs==1.13.5
requests==2.32.0
moviepy==2.1.2
pydub==0.25.1
Pillow==10.4.0
yt-dlp==2025.3.31
youtube-transcript-api==0.6.3
replicate==1.0.4
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.11.0
noisereduce==3.0.3
numpy==2.2.0
